% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_points_to_vcmm.R
\name{fit_points_to_vcmm}
\alias{fit_points_to_vcmm}
\title{label points by vcmm}
\usage{
fit_points_to_vcmm(x, margin_pars, M1, M2, F1, F2, cop1, cop2, mix)
}
\arguments{
\item{x}{data}

\item{margin_pars}{marginal parameters}

\item{M1}{vine structure for component 1}

\item{M2}{vine structure for component 2}

\item{F1}{copula families for component 1}

\item{F2}{copula families for component 2}

\item{cop1}{copula params for component 1}

\item{cop2}{copula params for component 2}

\item{mix}{mixing parameters for components}
}
\value{
labels
}
\description{
label points by vcmm
}
