% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_vineclust.R
\name{test_vineclust}
\alias{test_vineclust}
\title{fit model to toy example}
\usage{
test_vineclust(
  x,
  margin_pars,
  M1,
  M2,
  F1,
  F2,
  cop1,
  cop2,
  mix,
  iter = 100,
  maxit = 10
)
}
\arguments{
\item{x}{data}

\item{margin_pars}{marginal parameters}

\item{M1}{vine structure for component 1}

\item{M2}{vine structure for component 2}

\item{F1}{copula families for component 1}

\item{F2}{copula families for component 2}

\item{cop1}{copula params for component 1}

\item{cop2}{copula params for component 2}

\item{mix}{mixing parameters for components}

\item{iter}{iterations of ECM}

\item{maxit}{iterations for each CM step}
}
\value{
model
}
\description{
fit model to toy example
}
